chore: convert to mjs, update documentation
