+++ meteor/package.json
@@ -28,8 +28,8 @@
 		"lint:raw": "eslint --ext .ts --ext .js --ext .tsx --ext .jsx",
 		"lintfix": "yarn lint -- --fix",
 		"quickformat": "prettier \"__mocks__/**\" --write ; prettier \"lib/**\" --write ; prettier \"server/**\" --write ; prettier \"client/**\" --write ; prettier \"*.json\" --write ; prettier \"*.js\" --write ; prettier \"*.md\" --write",
-		"i18n-extract-pot": "node ./scripts/extract-i18next-pot.js -f \"{./client/ui/**/*.+(ts|tsx),./lib/mediaObjects.ts,./lib/rundownNotifications.ts,./server/migration/**/*.ts}\" -o i18n/template.pot",
-		"i18n-compile-json": "node ./scripts/i18n-compile-json.js",
+		"i18n-extract-pot": "node ./scripts/extract-i18next-pot.mjs -f \"{./client/ui/**/*.+(ts|tsx),./lib/mediaObjects.ts,./lib/rundownNotifications.ts,./server/migration/**/*.ts}\" -o i18n/template.pot",
+		"i18n-compile-json": "node ./scripts/i18n-compile-json.mjs",
 		"visualize": "meteor --production --extra-packages bundle-visualizer",
 		"release": "standard-version --commit-all",
 		"prepareChangelog": "yarn release -- --prerelease",
